Desktop/MergeMate/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.26.0
python-jose==3.3.0
openai==1.12.0
anthropic==0.18.1
python-gitlab==4.3.0
jira==3.5.2
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 